fixed some warnings
git-svn-id: svn://orca.physics.unc.edu/Orca@1440 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Misc/SynthesizeSingleton.h
@@ -87,7 +87,7 @@ return self; \
 return self; \
 } \
 \
-- (NSUInteger)retainCount \
+- (unsigned)retainCount \
 { \
 return NSUIntegerMax; \
 } \
